refactor (add interface) : 현재 이동정책은 랜덤값으로 이동하지만 언제든지 바뀔 수 있는 정책이라고 생각과 타입 유연성을 위해 이동정책을 인터페이스로 추출
